requirements.txt
appdirs==1.4.4
asgiref==3.3.1
black==20.8b1
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.4
django-environ==0.4.5
django-heroku==0.3.1
djangorestframework==3.12.2
environ==1.0
filelock==3.0.12
gunicorn==20.0.4
idna==2.10
import==0.0.1
mypy-extensions==0.4.3
pathspec==0.8.1
pipenv==2020.11.15
psycopg2==2.8.6
psycopg2-binary==2.8.6
python-dotenv==0.15.0
pytz==2020.4
regex==2020.11.13
requests==2.25.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.26.2
virtualenv==20.2.1
virtualenv-clone==0.5.4
whitenoise==5.2.0
